Add new Stories_Base_Controller class
