+++ includes/REST_API/Stories_Controller.php
@@ -27,21 +27,16 @@
 namespace Google\Web_Stories\REST_API;
 
 use Google\Web_Stories\Story_Post_Type;
-use stdClass;
 use WP_Query;
 use WP_Error;
 use WP_Post;
-use WP_REST_Posts_Controller;
 use WP_REST_Request;
 use WP_REST_Response;
 
 /**
- * Override the WP_REST_Posts_Controller class to add `post_content_filtered` to REST request.
- *
- * Class Stories_Controller
+ * Stories_Controller class.
  */
-class Stories_Controller extends WP_REST_Posts_Controller {
-
+class Stories_Controller extends Stories_Base_Controller {
 	const STYLE_PRESETS_OPTION = 'web_stories_style_presets';
 
 	/**
@@ -54,34 +49,6 @@ class Stories_Controller extends WP_REST_Posts_Controller {
 	];
 
 	const PUBLISHER_LOGOS_OPTION = 'web_stories_publisher_logos';
-	/**
-	 * Prepares a single story for create or update. Add post_content_filtered field to save/insert.
-	 *
-	 * @param WP_REST_Request $request Request object.
-	 *
-	 * @return stdClass|WP_Error Post object or WP_Error.
-	 */
-	protected function prepare_item_for_database( $request ) {
-		$prepared_story = parent::prepare_item_for_database( $request );
-
-		if ( is_wp_error( $prepared_story ) ) {
-			return $prepared_story;
-		}
-		// Ensure that content and story_data are updated together.
-		if (
-			( ! empty( $request['story_data'] ) && empty( $request['content'] ) ) ||
-			( ! empty( $request['content'] ) && empty( $request['story_data'] ) )
-		) {
-			return new WP_Error( 'rest_empty_content', __( 'content and story_data should always be updated together.', 'web-stories' ), [ 'status' => 412 ] );
-		}
-
-		// If the request is updating the content as well, let's make sure the JSON representation of the story is saved, too.
-		if ( isset( $request['story_data'] ) ) {
-			$prepared_story->post_content_filtered = wp_json_encode( $request['story_data'] );
-		}
-
-		return $prepared_story;
-	}
 
 	/**
 	 * Prepares a single story output for response. Add post_content_filtered field to output.
@@ -96,11 +63,6 @@ class Stories_Controller extends WP_REST_Posts_Controller {
 		$fields   = $this->get_fields_for_response( $request );
 		$data     = $response->get_data();
 		$schema   = $this->get_item_schema();
-
-		if ( in_array( 'story_data', $fields, true ) ) {
-			$post_story_data    = json_decode( $post->post_content_filtered, true );
-			$data['story_data'] = rest_sanitize_value_from_schema( $post_story_data, $schema['properties']['story_data'] );
-		}
 
 		if ( in_array( 'featured_media_url', $fields, true ) ) {
 			$image                      = get_the_post_thumbnail_url( $post, 'medium' );
@@ -175,14 +137,8 @@ class Stories_Controller extends WP_REST_Posts_Controller {
 		if ( $this->schema ) {
 			return $this->add_additional_fields_schema( $this->schema );
 		}
-		$schema = parent::get_item_schema();
 
-		$schema['properties']['story_data'] = [
-			'description' => __( 'Story data stored as a JSON object. Stored in post_content_filtered field.', 'web-stories' ),
-			'type'        => 'object',
-			'context'     => [ 'edit' ],
-			'default'     => [],
-		];
+		$schema = parent::get_item_schema();
 
 		$schema['properties']['featured_media_url'] = [
 			'description' => __( 'URL for the story\'s poster image (portrait)', 'web-stories' ),
